@@ -1,17 +1,20 @@
 module github.com/simonski/goutils
 
-go 1.14
+go 1.19
 
 require (
-	github.com/shirou/gopsutil v3.21.11+incompatible
+	github.com/shirou/gopsutil/v3 v3.22.7
 	github.com/simonski/cli v0.0.0-20220424130852-aaf734ce6ee0
 	golang.org/x/crypto v0.0.0-20220507011949-2cf3adece122
 	golang.org/x/term v0.0.0-20220411215600-e5f449aeb171
 )
 
 require (
-	github.com/stretchr/testify v1.7.0 // indirect
+	github.com/go-ole/go-ole v1.2.6 // indirect
+	github.com/lufia/plan9stats v0.0.0-20211012122336-39d0f177ccd0 // indirect
+	github.com/power-devops/perfstat v0.0.0-20210106213030-5aafc221ea8c // indirect
 	github.com/tklauser/go-sysconf v0.3.10 // indirect
+	github.com/tklauser/numcpus v0.4.0 // indirect
 	github.com/yusufpapurcu/wmi v1.2.2 // indirect
-	golang.org/x/sys v0.0.0-20220503163025-988cb79eb6c6 // indirect
+	golang.org/x/sys v0.0.0-20220520151302-bc2c85ada10a // indirect
 )
