@@ -25,12 +25,12 @@ func (a *SemVer) Eq(b *SemVer) bool {
 func (a *SemVer) Lt(b *SemVer) bool {
 	if a.Major < b.Major {
 		return true
-	} else if a.Major < b.Major {
+	} else if a.Major > b.Major {
 		return false
 	}
 	if a.Minor < b.Minor {
 		return true
-	} else if a.Minor < b.Minor {
+	} else if a.Minor > b.Minor {
 		return false
 	}
 	if a.Increment < b.Increment {
@@ -59,19 +59,19 @@ func (a *SemVer) Gt(b *SemVer) bool {
 func New(value string) (*SemVer, error) {
 	splits := strings.Split(value, ".")
 	if len(splits) != 3 {
-		return nil, errors.New("Not in format 'n.n.n'")
+		return nil, errors.New("not in format 'n.n.n'")
 	}
 	major, err := strconv.Atoi(splits[0])
 	if err != nil {
-		return nil, errors.New("Major version not an integer.")
+		return nil, errors.New("major version not an integer")
 	}
 	minor, err := strconv.Atoi(splits[1])
 	if err != nil {
-		return nil, errors.New("Minor version not an integer.")
+		return nil, errors.New("minor version not an integer")
 	}
 	increment, err := strconv.Atoi(splits[2])
 	if err != nil {
-		return nil, errors.New("Increment version not an integer.")
+		return nil, errors.New("increment version not an integer")
 	}
 	return &SemVer{Value: value, Major: major, Minor: minor, Increment: increment}, nil
 }
